refactor(src/components/NavBar/MobileNavBar/  src/components/NavBar/WebNavBar/) alteracao dos nomes da pasta @wander06 @stephanybrazeir0
